requirements.txt
astropy==5.1
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
cachetools==5.3.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
Counter==1.0.0
cryptography==37.0.4
cycler==0.11.0
filelock==3.12.2
Flask==2.1.3
fonttools==4.34.4
fsspec==2023.6.0
fuzzywuzzy==0.18.0
geographiclib==1.52
geopy==2.2.0
google-api-core==2.11.1
google-api-python-client==2.48.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
Google-Images-Search==1.4.6
googleapis-common-protos==1.60.0
googlesearch-python==1.2.3
h11==0.13.0
httplib2==0.22.0
huggingface-hub==0.16.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.4
Levenshtein==0.21.1
Markdown==3.4.4
MarkupSafe==2.1.1
matplotlib==3.5.2
mpmath==1.3.0
networkx==3.1
numpy==1.23.1
outcome==1.2.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
protobuf==4.24.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycraf==2.0.0
pyerfa==2.0.0.1
pyfiglet==0.8.post1
pyFirmata==1.1.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.21.1
python-resize-image==1.1.20
pytz==2022.1
PyYAML==6.0
rapidfuzz==3.2.0
regex==2023.6.3
requests==2.28.1
rsa==4.9
safetensors==0.3.2
scikit-learn==1.1.1
scipy==1.8.1
selenium==4.3.0
sgp4==2.21
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sympy==1.12
termcolor==1.1.0
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.1
transformers==4.31.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.10
Werkzeug==2.1.2
wsproto==1.1.0
